fix: an explicit null in config.json is a missing key, not a failure
Found by review of the merge-over-defaults decoder. A JSON null reached
the synthesized decoder, which has no default to fall back on and throws
on any field that is not Optional. ConfigStore treats a throw as an
unreadable file and sets the whole thing aside, so one null field cost
every setting, every workspace and every zone set.

The per-field decoder this replaced used decodeIfPresent, which reads a
null as absent and defaults that one field. withoutNulls restores that,
at any depth, before the merge.

A file can hold one for plenty of ordinary reasons: a hand edit, a field
an older release wrote as null, an exporter that spells "unset" that
way. None of them should cost the user their layouts.

The fix pushed ConfigStore.swift over the line limit, and it was holding
three types anyway. Split along them: Config, Config+Decoding,
LayoutItemSpec, and the store itself, which is now 58 lines.
