+++ App/Sources/plonk/Config+Decoding.swift
@@ -1,0 +1,82 @@
+import Foundation
+
+// Reading a config file.
+//
+// Swift synthesizes a decoder that demands every key and has no notion of a
+// property's default, so a tolerant read is not something the language will
+// write. Doing it per field cost a line each and made a forgotten line a
+// default that silently stopped applying. This does it once, in JSON, before
+// the decoder ever runs.
+
+extension Config {
+    /// Read a config file, tolerating every key it does not have.
+    ///
+    /// Swift synthesizes a decoder that demands every key, and there is no way
+    /// to tell it "use the property's default when one is missing". Spelling
+    /// that out by hand cost a line per field, and a new setting decoded only
+    /// if you remembered to add its line — which is a default that silently
+    /// stops applying, not a compiler error.
+    ///
+    /// So the file is merged over the defaults as JSON first. Every key is then
+    /// present, the synthesized decoder is enough, and adding a setting is the
+    /// field and nothing else.
+    static func decode(_ data: Data) throws -> Config {
+        guard let stored = try JSONSerialization.jsonObject(with: data) as? [String: Any] else {
+            throw Failure.notAnObject
+        }
+        let defaults = try JSONSerialization
+            .jsonObject(with: JSONEncoder().encode(Config())) as? [String: Any] ?? [:]
+        let merged = Self.merging(defaults, under: Self.withoutNulls(Self.migrated(stored)))
+        return try JSONDecoder().decode(
+            Config.self, from: JSONSerialization.data(withJSONObject: merged))
+    }
+
+    enum Failure: Error {
+        case notAnObject
+    }
+
+    /// `over` wins, key by key, recursing where both sides are objects. The
+    /// recursion is what lets a file name one field of `appearance` without
+    /// losing the rest of it.
+    private static func merging(_ base: [String: Any], under over: [String: Any]) -> [String: Any] {
+        base.merging(over) { mine, theirs in
+            guard let mine = mine as? [String: Any], let theirs = theirs as? [String: Any] else {
+                return theirs
+            }
+            return merging(mine, under: theirs)
+        }
+    }
+
+    /// Drops every explicit `null`, at any depth, so the default is what shows
+    /// through the merge.
+    ///
+    /// A file may hold one for any number of reasons: a hand edit, a field an
+    /// older release wrote as null, an exporter that spells "unset" that way.
+    /// Reaching the decoder it would be a hard failure on any field that is not
+    /// Optional, and a failure is not one setting reset — `load` sets the whole
+    /// file aside, which is every workspace and every zone set with it. So null
+    /// means absent here, which is what the per-field decoder this replaced did
+    /// with it.
+    private static func withoutNulls(_ value: [String: Any]) -> [String: Any] {
+        value.reduce(into: [String: Any]()) { result, pair in
+            switch pair.value {
+            case is NSNull:
+                return
+            case let nested as [String: Any]:
+                result[pair.key] = withoutNulls(nested)
+            default:
+                result[pair.key] = pair.value
+            }
+        }
+    }
+
+    /// Renames a key an older release wrote. Configs written before workspaces
+    /// could launch apps hold bare item arrays under "layouts".
+    private static func migrated(_ stored: [String: Any]) -> [String: Any] {
+        guard stored["workspaces"] == nil,
+              let layouts = stored["layouts"] as? [String: Any] else { return stored }
+        var stored = stored
+        stored["workspaces"] = layouts.mapValues { ["items": $0] }
+        return stored
+    }
+}
